arranging files to require fewer headers in library form
