+++ src/bmff.c
@@ -1,7 +1,10 @@
 #include <memory.h>
+#include <stdio.h>
 
 #include "bmff.h"
 #include "parse.h"
+#include "parse_common.h"
+#include "context.h"
 
 #define BOX_TYPE_IS(d,t) ((d)[0]==(t)[0] && (d)[1]==(t)[1] && (d)[2]==(t)[2] && (d)[3]==(t)[3])
 
@@ -55,7 +58,7 @@ size_t bmff_parse(BMFFContext *ctx, const uint8_t *data, size_t size, BMFFCode *
         // the correct order.
         uint32_t box_type = *((uint32_t*)(ptr+4));
 
-        printf("%c%c%c%c, size: %d\n", ptr[4], ptr[5], ptr[6], ptr[7], box_size);
+        //printf("%c%c%c%c, size: %d\n", ptr[4], ptr[5], ptr[6], ptr[7], box_size);
 
         int i=0;
         for(; i<13; ++i) {
@@ -68,9 +71,9 @@ size_t bmff_parse(BMFFContext *ctx, const uint8_t *data, size_t size, BMFFCode *
                 Box *box;
                 BMFFCode res = parse_map[i].parse_func(ctx, ptr, end-ptr, &box);
                 if(res == BMFF_OK) {
-                    printf("%c%c%c%c\n", box->type[0], box->type[1], box->type[2], box->type[3]);
+                    //printf("%c%c%c%c\n", box->type[0], box->type[1], box->type[2], box->type[3]);
                 } else {
-                    printf("Error paring box: %d\n", res);
+                    fprintf(stderr, "Error paring box: %d\n", res);
                 }
                 if(parser_is_container_type) {
                     bmff_context_alloc_stack_pop(ctx);
