+++ src/context.h
@@ -28,114 +28,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-#include "boxes.h"
-
-/**
- * Callback Event Ids.
- */
-typedef enum BMFFEventId {
-    BMFFEventFileType,
-    BMFFEventObjDescriptor,
-    BMFFEventMovieHeader,
-    BMFFEventTrackHeader,
-    BMFFEventMediaHeader,
-    BMFFEventVideoMediaHeader,
-    BMFFEventSoundMediaHeader,
-    BMFFEventSampleDescription,
-    BMFFEventVisualSample,
-    BMFFEventAudioSample,
-    BMFFEventHintSample,
-    BMFFEventSampleSize,
-    BMFFEventDataReference,
-    BMFFEventHandler,
-    BMFFEventTimeToSample,
-    BMFFEventSampleToChunk,
-    BMFFEventChunkOffset,
-    BMFFEventSyncSample,
-    BMFFEventMeta,
-} BMFFEventId;
-
-// forward declarations
-typedef struct BMFFContext BMFFContext;
-
-/**
- * Memory Allocator.
- * Allocate memory block.
- * See malloc C90 (C++98) definition.
- */
-typedef void * (*bmff_malloc) (size_t size);
-
-/**
- * Memory Allocator.
- * Allocate and zero initialize array.
- * See calloc C90 (C++98) definition.
- */
-typedef void * (*bmff_calloc) (size_t num, size_t size);
-
-/**
- * Memory Allocator.
- * Changes the size of the memory block pointed at by ptr.
- * See realloc C90 (C++98) definition.
- */
-typedef void * (*bmff_realloc) (void *ptr, size_t size);
-
-/**
- * Memory deallocator.
- */
-typedef void (*bmff_free) (void *mem);
-
-/**
- * Parse Callback.
- */
-typedef void (*bmff_on_event) (BMFFContext *ctx,
-                                BMFFEventId id,
-                                void *data);
-
-/**
- * Return codes.
- */
-typedef enum BMFFCode {
-    BMFF_OK                                 = 0x0000,
-    BMFF_INVALID_CONTEXT                    = 0x0001,
-    BMFF_INVALID_DATA                       = 0x0002,
-    BMFF_INVALID_SIZE                       = 0x0003,
-    BMFF_INVALID_PARAMETER                  = 0x0004,
-} BMFFCode;
-
-/**
- * LinkList of pointers.
- */
-typedef struct MemList {
-    size_t *addresses;
-    uint32_t count;
-    uint32_t used;
-    struct MemList *next;
-} MemList;
-
-/**
- * BMFF Parsing Context.
- */
-typedef struct BMFFContext {
-    bmff_malloc malloc;
-    bmff_realloc realloc;
-    bmff_calloc calloc;
-    bmff_free free;
-    // user specified callback that is called when Boxes are parsed
-    bmff_on_event callback;
-    // sample count set by the stsz or stz2 parser and used by the sdtp and stdp parsers.
-    uint32_t sample_count;
-    // current track sampler handler type used by the stsd and sgpd boxes to parse sample
-    // description data.
-    // this data comes from the active HandlerBox.
-    uint8_t handler_type[4];
-    // channel count set by the last AudioSampleEntry that was parsed.
-    // this is used by the chnl box parser.
-    uint32_t channel_count;
-    // version of the last sample description box. Used by the AudioSampleEntry box parser.
-    uint32_t sample_description_version;
-    // memory allocations stack (linked list)
-    MemList *allocs_stack;
-} BMFFContext;
+#include "bmff.h"
 
 /**
  * Adds a new to the stack of memory allocations.
