+++ src/descriptors.c
@@ -1,5 +1,4 @@
-#include "descriptors.h"
-#include "context.h"
+#include "bmff.h"
 #include "parse_common.h"
 
 /*
